pyproject.toml
[project]
name = "curve-dao"
version = "1.0.4"
description = "DAO governance tools for CurveDAO."
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.10"
license = {text = "AGPL-3.0-only License"}
keywords = ["ethereum", "smart-contracts", "evm", "vyper", "curve-finance", "dao", "defi"]
authors = [
    {name = "FiddyResearch", email = "fiddyresearch@gmail.com"},
]
maintainers = [
    {name = "FiddyResearch", email = "fiddyresearch@gmail.com"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

# Requirements
dependencies = [
    "curve-boa==0.1.10.1",
    "requests==2.32.3",
    "rich==13.8.0",
    "strenum==0.4.15",
]

[build-system]
requires = ["setuptools", "wheel"]

[tool.setuptools.packages.find]
include = ["curve_dao"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"

[project.urls]
Homepage = "https://pypi.org/project/curve-dao/"
GitHub = "https://github.com/bout3fiddy/curve-dao"
Changelog = "https://github.com/bout3fiddy/curve-dao/blob/main/CHANGELOG.md"
Issues = "https://github.com/bout3fiddy/curve-dao/issues"
